add example

+++ Design_pattern/PatternCompositeExample/Expression.cs
@@ -2,7 +2,7 @@
 {
     public abstract class Expression
     {
-        public abstract int Evalue();
+        public abstract double Evalue();
         public string Formate()
         {
             return string.Format("{0}={1}",ToString(), Evalue());
